sakila_python/pyproject.toml
[project]
name = "sakila-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=44.0.2",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.40",
]
